creando modelo, controlador, migracion, rutas y crud de Pnf

+++ api/routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\LoginController;
+use App\Http\Controllers\PnfController;
 use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -24,7 +25,14 @@ Route::middleware('auth:sanctum')->get('/users', [UserController::class, 'index'
 // Route::get('/users', [UserController::class, 'index']);
 Route::middleware('auth:sanctum')->post('/users', [UserController::class, 'store']);
 Route::middleware('auth:sanctum')->put('/users/{id}', [UserController::class, 'update']);
-Route::delete('/users/{id}', [UserController::class, 'destroy']);
+Route::middleware('auth:sanctum')->delete('/users/{id}', [UserController::class, 'destroy']);
+
+
+Route::middleware('auth:sanctum')->get('/pnf', [PnfController::class, 'index']);
+Route::middleware('auth:sanctum')->post('/pnf', [PnfController::class, 'store']);
+Route::middleware('auth:sanctum')->put('/pnf/{id}', [PnfController::class, 'update']);
+Route::middleware('auth:sanctum')->delete('/pnf/{id}', [PnfController::class, 'destroy']);
+
 
 Route::post('/login', [LoginController::class, 'authenticate']);
